ETL vocabulary: carry DRUG_STRENGTH through the generated scripts
00_vocabulary.sql never copied drug_strength, so the target CDM lost every
dosage row. Copy it in PART 5, gated on the reference actually holding the
table, filtered on all five concept columns (ingredient and unit ids are
foreign keys to concept too) and with the ATHENA date columns normalised —
a bare SELECT * fails the BIGINT -> DATE cast on INSERT.

99_prune_vocabulary.sql knew the table but pruned it wrong on two counts:
it kept rows on drug_concept_id alone, leaving keys pointing at ingredients
and units the same script had just deleted; and the keep-closure never
included those ingredients and units, since drug_strength describes them
while no CDM column references them — so the prune emptied the table for
every drug it had just kept. Both fixed, with a NULL guard on the closure:
the unit columns are nullable, and one NULL makes every
`NOT IN (SELECT ... FROM tmp_keep_concepts)` evaluate to NULL, silently
pruning nothing at all.

Also widen drug_strength's three value columns to DOUBLE. On DuckDB NUMERIC
means DECIMAL(18,3), and the ATHENA vocabulary ships numerator_value up to
2.26e15, which fails to load outright. The CDM spec types them as float;
only OHDSI's DuckDB DDL narrows them, so this diverges from upstream on
purpose and says so in place.

+++ apps/web/src/lib/schema-ddl/omop-5.4-ddl.ts
@@ -6,6 +6,9 @@
  * Primary keys are inline in CREATE TABLE statements.
  * Schema placeholder (@cdmDatabaseSchema) removed.
  * HINT comments removed (distribution hints for MPP engines, ignored by DuckDB).
+ *
+ * One deliberate divergence from upstream: drug_strength's three value columns
+ * are DOUBLE rather than NUMERIC — see the note on the table.
  */
 export const OMOP_54_DDL = `-- OMOP CDM v5.4 DDL (DuckDB)
 -- https://github.com/OHDSI/CommonDataModel
@@ -554,11 +557,16 @@ CREATE TABLE source_to_concept_map (
 CREATE TABLE drug_strength (
   drug_concept_id integer NOT NULL,
   ingredient_concept_id integer NOT NULL,
-  amount_value NUMERIC NULL,
+  -- DOUBLE, not the upstream NUMERIC: DuckDB reads NUMERIC as DECIMAL(18,3),
+  -- and the ATHENA vocabulary ships numerator_value up to 2.26e15 — loading it
+  -- into the upstream type fails outright ("Could not cast value
+  -- 2260000000000000.000000 to DECIMAL(18,3)"). The CDM spec itself types these
+  -- three as float; only the DuckDB DDL narrows them.
+  amount_value DOUBLE NULL,
   amount_unit_concept_id integer NULL,
-  numerator_value NUMERIC NULL,
+  numerator_value DOUBLE NULL,
   numerator_unit_concept_id integer NULL,
-  denominator_value NUMERIC NULL,
+  denominator_value DOUBLE NULL,
   denominator_unit_concept_id integer NULL,
   box_size integer NULL,
   valid_start_date date NOT NULL,
